remove some file
