--- src/modules/transfer/backend/m3db/aggregate.go
@@ -1,1 +1,0 @@
-package m3db
