--- src/modules/transfer/backend/m3db/util.go
@@ -1,1 +1,0 @@
-package m3db
